maintain a min. number of mixing-capable outbound peers
The minimum number of mixing-capable outbound peers is 3, unless a lower value
is configured for MaxPeers, in which case this lower value is used as the min.
number of mixing-capable outbound peers.

New outbound peers that do not support mixing will be disconnected if the number
of already connected peers is close to the target number of outbound peers (8 by
default) and there are not enough mixing-capable outbound peers yet.